refactor(ranker): split ranking helpers into focused modules
Move ranking orchestration, filters, quality scoring, source trust constants, and candidate variant selection into focused modules.

Keep whichllm.engine.ranker as a compatibility shim so existing public and test imports continue to work.

Refs #41
